TVB-1580 Fix problems when creating links for a group. Remove confusion and mix of op_group_id with dt_ids in link references.
git-svn-id: https://repo.thevirtualbrain.org/svn/tvb/trunk@6577 1c0e02f0-7929-43c0-8fb3-3293bf43b0d1

+++ framework_tvb/tvb/core/services/flow_service.py
@@ -350,7 +350,8 @@ class FlowService:
         Remove the link from the datatype given by dt_id to project given by project_id.
         """
         link = dao.get_link(dt_id, project_id)
-        dao.remove_entity(model.Links, link.id)
+        if link is not None:
+            dao.remove_entity(model.Links, link.id)
     
         
     def fire_operation(self, adapter_instance, current_user, project_id,
